blip/requirements.txt
apache_beam
mwparserfromhell
faiss-gpu
accelerate
datasets
numpy
transformer
tf-gpu
tensorflow
PIL
pandas
scikit-learn
matplotlib